requirements.txt
dnspython>=2.4.0
graphviz>=0.20.1
tldextract>=3.4.0
colorama>=0.4.6
pytest>=7.0.0
flake8>=6.0.0
